requirements.txt
asgiref==3.8.1
boto3==1.34.151
botocore==1.34.151
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==64.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
jmespath==1.0.1
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
